+++ src/components/core/status/StatusLight/StatusLight-types.ts
@@ -1,0 +1,7 @@
+export type Color = 'white' | 'black' | 'gray' | 'yellow' | 'red' | 'blue' | 'green';
+
+export interface ViewProps {
+  color: Color;
+  label?: string;
+  disabled?: boolean;
+}