+++ src/components/core/status/index.ts
@@ -1,1 +1,2 @@
 export { default as CircleLoader } from './CircleLoader';
+export { default as StatusLight } from './StatusLight';